move related services and models. improve paths
