+++ api/src/Blog/List.ts
@@ -1,8 +1,8 @@
 import { handlerFor } from "#api/lib/handler"
 import { S } from "#api/resources/lib"
-import { BlogPostView } from "#api/resources/views"
-import { BlogPostRepo } from "#api/services"
 import { Effect } from "effect-app"
+import { BlogPostView } from "./PostView.js"
+import { BlogPostRepo } from "./Repo.js"
 
 export class Request extends S.Req<Request>()("Blog.List", {}, {
   allowAnonymous: true,
