+++ api/src/Blog/Publish.ts
@@ -2,12 +2,13 @@ import { handlerFor } from "#api/lib/handler"
 import { OperationsDefault } from "#api/lib/layers"
 import { BogusEvent } from "#api/resources/Events"
 import { S } from "#api/resources/lib"
-import { BlogPostRepo, Events, Operations } from "#api/services"
-import { BlogPostId } from "#models/Blog"
+import { Events, Operations } from "#api/services"
 import { Duration, Effect, Schedule } from "effect-app"
 import { NotFoundError } from "effect-app/client"
 import { OperationId } from "effect-app/Operations"
 import { NonEmptyString2k, NonNegativeInt } from "effect-app/Schema"
+import { BlogPostId } from "./models.js"
+import { BlogPostRepo } from "./Repo.js"
 
 export class Request extends S.Req<Request>()("Blog.PublishPost", {
   id: BlogPostId
