+++ api/src/Blog/models.ts
@@ -1,5 +1,5 @@
+import { UserFromId } from "#Accounts/User"
 import { S } from "effect-app"
-import { UserFromId } from "./User.js"
 
 export const BlogPostId = S.prefixedStringId<BlogPostId>()("post", "BlogPostId")
 export interface BlogPostIdBrand {
