+++ api/src/HelloWorld.controllers.ts
@@ -1,11 +1,11 @@
 import { matchFor, Router } from "#api/lib/routing"
-import { UserRepo } from "#api/services"
-import { User } from "#models/User"
 import { HelloWorldRsc } from "#resources"
 import { GetHelloWorld } from "#resources/HelloWorld"
 import { getRequestContext } from "@effect-app/infra/api/setupRequest"
 import { generate } from "@effect-app/infra/test"
 import { Effect, S } from "effect-app"
+import { User } from "./Accounts/User.js"
+import { UserRepo } from "./Accounts/UserRepo.js"
 
 export default Router(HelloWorldRsc)({
   dependencies: [UserRepo.Default],
