+++ api/src/resources/HelloWorld.ts
@@ -1,6 +1,6 @@
+import { UserView } from "#Accounts/UserView"
 import { RequestContext } from "@effect-app/infra/RequestContext"
 import { S } from "./lib.js"
-import { UserView } from "./views.js"
 
 class Response extends S.Class<Response>()({
   now: S.Date.withDefault,
