+++ api/src/resources/views.ts
@@ -1,4 +1,3 @@
 // codegen:start {preset: barrel, include: ./views/*.ts}
-export * from "./views/PostView.js"
-export * from "./views/UserView.js"
+
 // codegen:end
