--- api/src/services/DBContext.ts
@@ -1,4 +1,0 @@
-// codegen:start {preset: barrel, include: ./DBContext/* }
-export * from "./DBContext/BlogPostRepo.js"
-export * from "./DBContext/UserRepo.js"
-// codegen:end
