+++ api/tsconfig.json
@@ -10,6 +10,12 @@
       "#models": [
         "./src/models.js"
       ],
+      "#Accounts/*": [
+        "./src/Accounts/*.js"
+      ],
+      "#Blog/*": [
+        "./src/Blog/*.js"
+      ],
       "#resources/*": [
         "./src/resources/*.js"
       ],
